requirements.txt
Django==5.2.6
Pillow==10.4.0
mysqlclient==2.2.7
python-dotenv==1.2.1
stripe==12.1.0

